Move OpenStax-related handler stuff to own package
